geometry: move pullback metric to the metric header
