+++ lib/mito/geometry/public.h
@@ -30,9 +30,6 @@
 #include "polar.h"
 #include "spherical.h"
 
-// induced metric (submanifolds)
-#include "induced_metric/public.h"
-
 // classes implementation
 #include "CoordinateSystem.h"
 #include "Point.h"
