--- lib/mito/geometry/pullback_metric.h
@@ -1,14 +1,0 @@
-// -*- c++ -*-
-//
-// Copyright (c) 2020-2026, the MiTo Authors, all rights reserved
-//
-
-// code guard
-#pragma once
-
-
-// publish the pullback metric interface (replaces induced metric)
-#include "pullback_metric/public.h"
-
-
-// end of file
