--- lib/mito/geometry/pullback_metric/forward.h
@@ -1,19 +1,0 @@
-// -*- c++ -*-
-//
-// Copyright (c) 2020-2026, the MiTo Authors, all rights reserved
-//
-
-// code guard
-#pragma once
-
-
-namespace mito::geometry {
-
-    // primary template: pullback of a metric along a parametrization
-    template <class...>
-    struct pullback_metric {};
-
-}
-
-
-// end of file
