--- lib/mito/geometry/pullback_metric/public.h
@@ -1,17 +1,0 @@
-// -*- c++ -*-
-//
-// Copyright (c) 2020-2026, the MiTo Authors, all rights reserved
-//
-
-// code guard
-#pragma once
-
-
-// get the forward declarations
-#include "forward.h"
-
-// get the pullback metric specialization
-#include "metric.h"
-
-
-// end of file
